requirements.txt
requests
tqdm
PyYAML
onnxruntime
soundfile